requirements_ml.txt
# Machine Learning Dependencies
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Optional: For plotting training curves
matplotlib>=3.7.0
